+++ apps/web/src/components/Composer/Actions/CollectSettings/AmountConfig.tsx
@@ -5,7 +5,7 @@ import type { Erc20 } from '@hey/lens';
 import { OpenActionModuleType } from '@hey/lens';
 import { Input } from '@hey/ui';
 import { type FC } from 'react';
-import { useCollectModuleStore } from 'src/store/useCollectModuleStore';
+import { useCollectModuleStore } from 'src/store/non-persisted/useCollectModuleStore';
 
 interface AmountConfigProps {
   enabledModuleCurrencies?: Erc20[];
