+++ apps/web/src/components/Composer/Actions/CollectSettings/CollectLimitConfig.tsx
@@ -2,7 +2,7 @@ import ToggleWithHelper from '@components/Shared/ToggleWithHelper';
 import { StarIcon } from '@heroicons/react/24/outline';
 import { Input } from '@hey/ui';
 import { type FC } from 'react';
-import { useCollectModuleStore } from 'src/store/useCollectModuleStore';
+import { useCollectModuleStore } from 'src/store/non-persisted/useCollectModuleStore';
 
 interface CollectLimitConfigProps {
   setCollectType: (data: any) => void;
