+++ apps/web/src/components/Composer/Actions/CollectSettings/TimeLimitConfig.tsx
@@ -4,7 +4,7 @@ import type { CollectModuleType } from '@hey/types/hey';
 import { Input } from '@hey/ui';
 import { getNumberOfDaysFromDate, getTimeAddedNDay } from '@lib/formatTime';
 import { type FC } from 'react';
-import { useCollectModuleStore } from 'src/store/useCollectModuleStore';
+import { useCollectModuleStore } from 'src/store/non-persisted/useCollectModuleStore';
 
 interface TimeLimitConfigProps {
   setCollectType: (data: CollectModuleType) => void;
