+++ apps/web/src/components/Home/RecommendedProfiles.tsx
@@ -15,7 +15,7 @@ import { motion } from 'framer-motion';
 import type { FC } from 'react';
 import { useState } from 'react';
 import useProfileStore from 'src/store/persisted/useProfileStore';
-import { useTimelineStore } from 'src/store/useTimelineStore';
+import { useTimelineStore } from 'src/store/non-persisted/useTimelineStore';
 
 import Suggested from './Suggested';
